refactor(frontend): BacktestData 페이지 데이터 흐름 컨벤션 준수 (api → hooks → pages)
- types/data.ts 신설: Dataset, DataType, StorageInfo 타입 분리
- hooks/useBacktestData.ts 신설: useDatasets, useStorageInfo, useDeleteDataset
- api/data.ts: 타입을 types/data.ts에서 re-export
- pages/BacktestData.tsx: api/ 직접 import 제거, hooks 경유로 전환

Refs #871
