+++ frontend/src/types/data.ts
@@ -1,0 +1,17 @@
+export type DataType = 'ohlcv' | 'fundamental'
+
+export interface Dataset {
+  id: string
+  symbol: string
+  timeframe: string
+  data_type: DataType
+  start_date: string
+  end_date: string
+  row_count: number
+}
+
+export interface StorageInfo {
+  total_bytes: number
+  total_mb: number
+  by_timeframe: Record<string, number>
+}
